[MAT-187] Moved exceptions
